+++ scripts/dice.mjs
@@ -1,5 +1,8 @@
 export async function abilityCheck({ ability, label, actor }) {
-  const { difficulty, resources } = await _getCheckOptions('CF.Rolls.AbilityCheck.Title', actor.system.attributes.resource);
+  const {
+    difficulty,
+    resources,
+  } = await _getCheckOptions('CF.Rolls.AbilityCheck.Title', actor.system.attributes.resource);
 
   await _doCheck({
     dice: ability,
@@ -12,7 +15,10 @@ export async function abilityCheck({ ability, label, actor }) {
 
 export async function specialtyCheck(specialty) {
   const actor = specialty.actor;
-  const { difficulty, resources } = await _getCheckOptions('CF.Rolls.SpecialtyCheck.Title', actor.system.attributes.resource);
+  const {
+    difficulty,
+    resources,
+  } = await _getCheckOptions('CF.Rolls.SpecialtyCheck.Title', actor.system.attributes.resource);
 
   await _doCheck({
     dice: actor.system.abilities[specialty.system.ability],
@@ -30,7 +36,10 @@ export async function specialtyCheck(specialty) {
 
 export async function weaponCheck(weapon) {
   const actor = weapon.actor;
-  const { difficulty, resources } = await _getCheckOptions('CF.Rolls.WeaponCheck.Title', actor.system.attributes.resource);
+  const {
+    difficulty,
+    resources,
+  } = await _getCheckOptions('CF.Rolls.WeaponCheck.Title', actor.system.attributes.resource);
 
   await _doCheck({
     dice: actor.system.abilities[weapon.system.ability],
@@ -42,7 +51,7 @@ export async function weaponCheck(weapon) {
   });
 }
 
-export async function reroll({ actor, available, bonus, difficulty, label, type, usable }) {
+export async function reroll({ actor, available, bonus, difficulty, label, type, usable, message }) {
   const i18nKey = 'specialty' === type ? 'CF.Rolls.SpecialtyCheck.Card.Title' : 'CF.Rolls.Damages.Card.Title';
 
   await _doCheck({
@@ -56,6 +65,7 @@ export async function reroll({ actor, available, bonus, difficulty, label, type,
     actor,
     bonus,
     difficulty,
+    message,
   });
 }
 
@@ -88,7 +98,7 @@ export async function useWeapon({ actor, dice, label, reroll }) {
   await _createChatMessage(actor, rollResult, chatContent, CONST.CHAT_MESSAGE_TYPES.ROLL);
 }
 
-async function _doCheck({ actor, bonus, dice, difficulty, reroll, title, usedResources, weapon }) {
+async function _doCheck({ actor, bonus, dice, difficulty, reroll, title, usedResources, weapon, message }) {
   if (usedResources) {
     // Ensure to not use more resources than necessary
     usedResources = Math.min(difficulty, usedResources);
@@ -103,6 +113,7 @@ async function _doCheck({ actor, bonus, dice, difficulty, reroll, title, usedRes
       difficulty,
       title,
       weapon,
+      message,
     });
 
     await _handleRollResult({ canReroll: false, actor, difficulty, usedResources });
@@ -157,7 +168,12 @@ async function _doCheck({ actor, bonus, dice, difficulty, reroll, title, usedRes
     weapon,
   });
 
-  await _createChatMessage(actor, rollResult, chatContent, CONST.CHAT_MESSAGE_TYPES.ROLL);
+  if (message) {
+    await message.update({ content: chatContent });
+  } else {
+    await _createChatMessage(actor, rollResult, chatContent, CONST.CHAT_MESSAGE_TYPES.ROLL);
+  }
+
   await _handleRollResult({ actor, canReroll: rerollData.can, difficulty, rollResult, usedResources });
 }
 
@@ -170,7 +186,7 @@ function _getRollResult(dice, bonus) {
   return new Roll(formula).roll({ async: true });
 }
 
-async function _rollNoRoll({ title, actor, difficulty, weapon }) {
+async function _rollNoRoll({ title, actor, difficulty, weapon, message }) {
   const chatContent = await renderTemplate('systems/channel-fear/templates/partials/roll/roll-card.hbs', {
     actorId: actor.id,
     failure: false,
@@ -181,14 +197,18 @@ async function _rollNoRoll({ title, actor, difficulty, weapon }) {
     weapon,
   });
 
-  _createChatMessage(actor, null, chatContent, CONST.CHAT_MESSAGE_TYPES.OTHER);
+  if (message) {
+    await message.update({ content: chatContent });
+  } else {
+    _createChatMessage(actor, null, chatContent, CONST.CHAT_MESSAGE_TYPES.OTHER);
+  }
 }
 
 function _createChatMessage(actor, rollResult, content, type) {
   return ChatMessage.create({
     user: game.user.id,
     speaker: ChatMessage.getSpeaker({ actor }),
-    roll: rollResult,
+    rolls: rollResult,
     content: content,
     sound: CONFIG.sounds.dice,
     type: type,